requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.0
pydantic==1.10.12
